requirements.txt
aiohappyeyeballs==2.3.5
aiohttp==3.10.3
aiosignal==1.3.1
APScheduler==3.10.4
attrs==24.2.0
backports.tarfile==1.2.0
black==24.8.0
build==1.2.1
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
docutils==0.21.2
frozenlist==1.4.1
idna==3.7
importlib_metadata==8.2.0
jaraco.classes==3.4.0
jaraco.context==5.3.0
jaraco.functools==4.0.2
keyring==25.3.0
markdown-it-py==3.0.0
mdurl==0.1.2
more-itertools==10.4.0
multidict==6.0.5
mypy-extensions==1.0.0
nh3==0.2.18
packaging==24.1
pathspec==0.12.1
pkginfo==1.10.0
platformdirs==4.2.2
Pygments==2.18.0
pyproject_hooks==1.1.0
python-dotenv==1.0.1
pytz==2024.1
readme_renderer==44.0
redis==4.6.0
requests==2.32.3
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.7.1
signalbot==0.10.1
six==1.16.0
twine==5.1.1
tzlocal==5.2
urllib3==2.2.2
websockets==10.4
yarl==1.9.4
zipp==3.20.0
